BACKPORT: menu: Ignore prompt variable if timeout is != 0
Since 739e8361f3fe78038251216df6096a32bc2d5839, a system with the following
/boot/extlinux/extlinux.conf (which sets timeout to 50) immediately boots the
first entry in the config without displaying a boot menu.  According to the
description, that should only happen if both prompt and timeout are set to zero
in the config, but it also happens with timeout set to a non-zero value.

Reported-by: Karsten Merker <merker@debian.org>
Signed-off-by: Manuel Traut <manuel.traut@mt.com>
Change-Id: I176a2987ad73936ba3a07078b501b353abfa6fe7
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>