common: image-fdt: Add support to dynamically set root device
This allows us to directly set rootdevice from u-boot itself, By doing so, we
ensure that U-Boot consistently selects the desired root device, even in
scenarios where multiple storage interfaces are simultaneously connected to
the board.

For instance, in cases where a board is equipped with eMMC, NVMe, and SD-Card
interfaces, all containing the same system with identical root PARTUUID, this
update ensures that the root device is chosen based on the priority established
within the function.

Specifically, it follows the order of preference:
NVMe first, followed by SD-Card, and finally eMMC.

Change-Id: I4d8ccc937ec0af2b6db2f4f0c08915e45f4a5dfb
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>